TestMathBits: test with neg and conj bit set
The case where both the negative and conjugate bits are set
isn't tested currently despite being handled explicitly by `copy`.
In theory this shouldn't matter because neg_bit is only used for real
values, but it does mean the code in copy is untested. So, this just
runs it with a single sample as a sanity check.

ghstack-source-id: 70222a8a215977eac3f50ea8bbf3d61ac136d9ef
Pull Request resolved: https://github.com/pytorch/pytorch/pull/68948