Provide functionality to write custom ShardedTensor ops.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/69874

We have a handful of ops supported for ShardedTensor via
``__torch_function__`` dispatch. However, we currently can't cover all torch
operators and having a way for users to extend this functionality will make
this functionality much more general.

In this PR, I've introduced a @custom_sharded_op decorator which can be used to
register a custom sharded op implementation.
ghstack-source-id: 145809101

Differential Revision: [D33078587](https://our.internmc.facebook.com/intern/diff/D33078587/)